+++ CHANGELOG.md
@@ -1,11 +1,11 @@
 
 1.2.0 -> 2.0.1:
 * Minor bugfix on memory checker
-* Update external software used
 
 1.2.0 -> 2.0.0:
 * Update the wepsim engine to 1.8.1
 * To use the native checker support from 1.8.1
+* Update external software used
 
 1.1.0 -> 1.2.0:
 * Update the wepsim engine to 1.7.6